Moved oarsing functions to gomodel/pathtemplateparser.go
